% Generated by roxygen2 (4.0.2): do not edit by hand
\name{looptimer}
\alias{looptimer}
\title{Loop timer}
\usage{
looptimer(tim, n, i, when_ready = FALSE)
}
\arguments{
\item{tim}{Previous instance of timer to be updated}

\item{n}{total count in the loop}

\item{i}{iteration that was just finished}

\item{when_ready}{if true add ETA to message.}
}
\description{
For timing loops
}
\details{
Approximative iteration speed, time left, and ETA.
}
\examples{
N <- 100
t0 <- timer()
for(i in 1:N){
 # do something
 Sys.sleep(runif(1))
 t0 <- timer(t0, N, i, TRUE)
 cat(t0$message,"\\n")
}
}

