% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_block.R
\name{split_block}
\alias{split_block}
\title{Split a vector to blocks of certain size}
\usage{
split_block(i, n)
}
\arguments{
\item{i}{vector to split}

\item{n}{elements of i per block}
}
\description{
Split a vector to blocks of certain size
}
\details{
last block might not contain n elements, but thats ok.
}
