% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.looptimer}
\alias{summary.looptimer}
\title{Summary method for looptimer}
\usage{
\method{summary}{looptimer}(x, ...)
}
\arguments{
\item{x}{looptimer-object}
}
\description{
Summary method for looptimer
}
\details{
Prints the iterations, average time and the total time taken. Perhaps call after the loop.
}
