% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/looptimer.R
\name{looptimer}
\alias{looptimer}
\title{Loop timer}
\usage{
looptimer(tim, n, i, when_ready = TRUE, memory = 50, prefix = "")
}
\arguments{
\item{tim}{Previous instance of timer to be updated}

\item{n}{total count in the loop}

\item{i}{iteration that was just finished}

\item{when_ready}{if true add ETA to message.}

\item{memory}{How many iteration back remember when computing the average time}

\item{prefix}{prefix for print message}
}
\description{
For timing loops
}
\details{
Approximative iteration speed, time left, and ETA.
}
\examples{


N <- 100
t0 <- looptimer()
for(i in 1:N){
 # do something
 Sys.sleep(runif(1))
 print(  t0 <- looptimer(t0, N, i, TRUE) )
}


}
