% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/looptimer.R
\name{looptimer}
\alias{looptimer}
\title{Loop timer}
\usage{
looptimer(
  tim,
  n,
  i,
  when_ready = TRUE,
  memory = 50,
  prefix = "",
  endline = "\\n",
  printevery = 1,
  fg = NULL,
  bg = NULL,
  use_msg = TRUE
)
}
\arguments{
\item{tim}{Previous instance of timer to be updated}

\item{n}{total count in the loop}

\item{i}{iteration that was just finished}

\item{when_ready}{if true add ETA to message.}

\item{memory}{How many iteration back remember when computing the average time}

\item{prefix}{prefix for print message}

\item{endline}{If print called, which endline character to use? Default '\\n'. Usual alternative is something like '   \\r'}

\item{printevery}{Print every nth iteration only. Affects print-method.}

\item{fg}{Use ANSI color for the print text? Integer from 1 to 256.}

\item{bg}{Use ANSI color for the print background? Integer from 1 to 256.}

\item{use_msg}{Use message for printing? Default is TRUE. If not, use cat}
}
\description{
For timing loops
}
\details{
Approximative iteration speed, time left, and ETA.
}
\examples{


N <- 100
t0 <- looptimer(n = N, fg = 2, prefix = "[testing]", printevery = 10)
for(i in 1:N){
 # do something
 Sys.sleep(runif(1))
 print(  t0 <- looptimer(t0) )
}


}
